FCBHDBP-363 It has moved the StageC-specific classes to new folder: stagec. It has renamed the file: LanguageRecordStageC.py to StageCLanguageRecord.py.
